Added all the classes in a header file and made changes in the code accordingly
