+++ NumberGuessGame/main.cpp
@@ -1,37 +1,4 @@
-#include<iostream>
-#include "easy.cpp"
-#include "hard.cpp"
-using namespace std;
-
-class NumberGuessGame : virtual EasyGame, virtual HardGame
-{
-    private:
-
-        int c;
-
-    public:
-
-        void play()
-        {
-            cout<<"1. Hard mode.\n";
-            cout<<"2. Easy mode.\n";
-            cout<<"Enter your choice:\n";
-            cin>>c;
-
-            switch(c)
-            {
-                case 1:
-                    HardGame :: play();
-                break;
-                case 2:
-                    EasyGame :: play();
-                break;
-                default:
-                    cout<<"Wrong choice!\n";
-                    exit(0);
-            }
-        }
-};
+#include "gameplay.cpp"
 
 int main()
 {
@@ -41,7 +8,6 @@ int main()
     {
         ob.play();
     }
-    
 
     return 0;
 }
